Add files via upload
hello everyone! I have an issue. Every time a play in my minecraft world and kill a few mobs to test the mob loot-tables i changed a crash appears. It is really strange because it happens out of nowhere and everything is doing fine at the start until i kill some mobs. Sorry for my bad english. Here is the Crash Report and my Crafttweaker script